code for simulations
this file contains all the codes for training & testing data, building the model, getting the wining rate between every 2 team (when home=1 or 0, and the simulation results. Note: the wining rate can be changed after every round, but for the baseline, I kept them same for all the four rounds.